+++ packages/client/src/layouts/CoverLayout.vue
@@ -23,7 +23,6 @@ const { cssVars: albumChrome } = useAlbumTheme({
   artworkUrl: () => props.artworkUrl,
   track: () => props.track,
   progress: () => props.progress,
-  includeBackground: true,
 });
 
 const backgroundRef = computed(() => props.background);
@@ -69,10 +68,10 @@ const layoutClass = computed(() => ({
 }));
 
 const rootStyle = computed(() => {
-  if (props.background === 'gradient-simple') {
-    return albumChrome.value;
+  if (usesDynamicBackground.value) {
+    return { ...albumChrome.value };
   }
-  return backgroundStyle.value || {};
+  return { ...albumChrome.value, ...(backgroundStyle.value || {}) };
 });
 </script>